@@ -1,6 +1,9 @@
 package com.example.booking.exception;
 
 import com.example.booking.util.RestResponse;
+import feign.FeignException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -15,6 +18,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @RestControllerAdvice
+@Slf4j
 public class GlobalException {
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
@@ -23,7 +27,7 @@ public class GlobalException {
         BindingResult result = ex.getBindingResult();
         List<FieldError> fieldErrors = result.getFieldErrors();
 
-        List<String> errors = fieldErrors.stream().map(fe -> fe.getDefaultMessage()).collect(Collectors.toList());
+        List<String> errors = fieldErrors.stream().map(DefaultMessageSourceResolvable::getDefaultMessage).collect(Collectors.toList());
 
         RestResponse<Object> response = new RestResponse<>();
         response.setStatus("Error");
@@ -35,8 +39,8 @@ public class GlobalException {
 
     }
 
-    @ExceptionHandler(value = { UsernameNotFoundException.class, BadCredentialsException.class })
-    public ResponseEntity<RestResponse<Object>> handleException(Exception ex) {
+    @ExceptionHandler(value = {UsernameNotFoundException.class, BadCredentialsException.class})
+    public ResponseEntity<RestResponse<Object>> handleLoginException(Exception ex) {
         System.out.println("Ex UsernameNotFoundException BadCredentialsException");
 
         RestResponse<Object> response = new RestResponse<Object>();
@@ -48,8 +52,23 @@ public class GlobalException {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
 
     }
+
+    @ExceptionHandler(FeignException.class)
+    public ResponseEntity<RestResponse<Object>> handleFeignException(FeignException ex) {
+        System.out.println("Ex handleFeignException");
+        RestResponse<Object> response = RestResponse.builder().status("Error").build();
+
+        response.setError(ex.getMessage());
+        response.setStatusCode(ex.status());
+        response.setMessage(ex.getMessage());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+
+    }
+
     @ExceptionHandler(Exception.class)
     public ResponseEntity<RestResponse<Object>> handleCommonException(Exception ex) {
+            log.error("Exception: {}", ex);
         System.out.println("Ex common");
 //    ex.printStackTrace();
         RestResponse<Object> response = new RestResponse<Object>();
