@@ -8,29 +8,31 @@ import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 @EnableWebSecurity
-@EnableMethodSecurity(securedEnabled = true, jsr250Enabled = true, prePostEnabled = true)
+@EnableMethodSecurity
 public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http, AuthenticationEntryPoint authenticationEntryPoint) throws Exception {
         final String[] PUBLIC_ENPOINT = {
-                "/auth/login",
+                "/bookings/**",
+                "/auth/register",
                 "/auth/exchange-token"
         };
 
-        http.csrf(c -> c.disable());
+        http.csrf(AbstractHttpConfigurer::disable);
 
         http.authorizeHttpRequests(auth -> auth.requestMatchers(PUBLIC_ENPOINT)
                 .permitAll()
                 .anyRequest().authenticated());
 
-        http.formLogin(f -> f.disable());
+        http.formLogin(AbstractHttpConfigurer::disable);
 
         http.oauth2ResourceServer(oauth -> oauth.jwt(Customizer.withDefaults())
                 .authenticationEntryPoint(authenticationEntryPoint));
