@@ -1,28 +1,40 @@
 package com.example.booking.entity;
 
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @Entity
 @Table(name = "bookings")
 public class Booking extends SuperEntity  {
+    @Column(name = "start_date")
+    private LocalDate startDate;
+    @Column(name = "end_date")
+    private LocalDate endDate;
     @Column(name = "start_time")
-    private LocalDateTime startTime;
+    private LocalTime startTime;
     @Column(name = "end_time")
-    private LocalDateTime endTime;
+    private LocalTime endTime;
     @Column(name = "total_price")
     private Double totalPrice;
+    @Column
+    @Enumerated(EnumType.STRING)
     private BookingStatus status;
 
     @ManyToOne
     @JoinColumn(name = "sport_field_id")
+    @JsonManagedReference
     private SportField sportField;
 
 }
