@@ -1,30 +1,38 @@
 package com.example.booking.entity;
 
 import com.example.booking.config.generator.GenerateID;
+import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.GeneratedColumn;
 import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.UpdateTimestamp;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 import java.time.LocalDateTime;
+import java.util.Date;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @MappedSuperclass
+@EqualsAndHashCode(of = "id")
 public abstract class SuperEntity {
     @Id
-    @GeneratedValue(generator = "GenerateID")
-    @GenericGenerator(name = "GenerateID", type = GenerateID.class)
+//    @GeneratedValue(generator = "GenerateID")
+//    @GenericGenerator(name = "GenerateID", type = GenerateID.class)
     private String id;
-    @CreatedDate
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false)
+    private Date createdAt;
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    private Date updatedAt;
 }
